Optimizations: done
fix #3

+++ src/resources/optimizations.js
@@ -27,17 +27,17 @@ class Optimizations {
 	 * @see {@link https://route4me.io/docs/#create-an-optimization Route4Me API}
 	 * @category Optimizations
 	 *
-	 * @param  {jsonschema:Optimizations.CreateRequest} params Parameters for new optimization
+	 * @param  {jsonschema:Optimizations.CreateRequest} optimization Parameters for new optimization
 	 * @param  {module:route4me-node~RequestCallback<jsonschema:Optimizations.Response>}  [callback]
 	 */
-	create(params, callback) {
-		const verror = this.r._validate(params, "Optimizations.CreateRequest")
-		if (verror) { return callback(verror) }
+	create(optimization, callback) {
+		const verror = this.r._validateArgs(arguments, "sdk:Optimizations.create")
+		if (verror instanceof Error) { return callback(verror) }
 
 		return this.r._makeRequest({
 			method: "POST",
 			path: "/api.v4/optimization_problem.php",
-			body: params,
+			body: optimization,
 			schemaName: "Optimizations.Response",
 		}, callback)
 	}
@@ -53,6 +53,9 @@ class Optimizations {
 	 * @param  {module:route4me-node~RequestCallback<jsonschema:Optimizations.Response>} [callback]
 	 */
 	get(id, callback) {
+		const verror = this.r._validateArgs(arguments, "sdk:Optimizations.get")
+		if (verror instanceof Error) { return callback(verror) }
+
 		if (typeof id !== "string" && typeof id !== "number") {
 			return callback(new errors.Route4MeError("'id' parameter MUST be a string"))
 		}
@@ -79,6 +82,9 @@ class Optimizations {
 	 * @param {module:route4me-node~RequestCallback<jsonschema:Optimizations.ResponseMany>} [callback]
 	 */
 	list(states, limit, offset, callback) {
+		const verror = this.r._validateArgs(arguments, "sdk:Optimizations.list")
+		if (verror instanceof Error) { return callback(verror) }
+
 		const _states = utils.parseStates(states)
 		if (_states instanceof Error) {
 			return callback(_states)
@@ -107,13 +113,13 @@ class Optimizations {
 	 * @since 0.1.7
 	 *
 	 * @param {(integer|string)} id Optimization Problem ID `optimization_problem_id`
-	 * @param {jsonschema:Optimizations.CreateRequest}   data New values for `Optimization`
+	 * @param {jsonschema:Optimizations.CreateRequest}   optimization - New values for `Optimization`
 	 * @param {boolean} reoptimize Determine, whether the `Optimization` should be reoptimized
 	 * @param {module:route4me-node~RequestCallback<jsonschema:Optimizations.Response>} [callback]
 	 */
-	update(id, data, reoptimize, callback) {
-		const verror = this.r._validate(data, "Optimizations.CreateRequest")
-		if (verror) { return callback(verror) }
+	update(id, optimization, reoptimize, callback) {
+		const verror = this.r._validateArgs(arguments, "sdk:Optimizations.update")
+		if (verror instanceof Error) { return callback(verror) }
 
 		return this.r._makeRequest({
 			method: "PUT",
@@ -122,11 +128,10 @@ class Optimizations {
 				"optimization_problem_id": id,
 				"reoptimize": reoptimize ? "1" : "0",
 			},
-			body: data,
+			body: optimization,
 			schemaName: "Optimizations.Response",
 		}, callback)
 	}
-
 
 	/**
 	 * Remove an existing optimization belonging to an user.
@@ -134,6 +139,8 @@ class Optimizations {
 	 * @see {@link https://route4me.io/docs/#remove-an-optimization  Route4Me API}
 	 * @category Optimizations
 	 * @since 0.1.7
+	 *
+	 * @todo TODO: There is no schema for validation an output
 	 *
 	 * @example
 	 * const response = {
@@ -146,6 +153,9 @@ class Optimizations {
 	 *     [callback]
 	 */
 	remove(id, callback) {
+		const verror = this.r._validateArgs(arguments, "sdk:Optimizations.remove")
+		if (verror instanceof Error) { return callback(verror) }
+
 		return this.r._makeRequest({
 			method: "DELETE",
 			path: "/api.v4/optimization_problem.php",
@@ -155,6 +165,70 @@ class Optimizations {
 			schemaName: "Optimizations.Response.Remove",
 		}, callback)
 	}
+
+	/**
+	 * Insert an address into an optimization, resulting in the recalculation of optimal routes.
+	 *
+	 * @see {@link https://route4me.io/docs/#insert-an-address-into-an-optimization Route4Me API}
+	 * @category Optimizations
+	 * @since 0.1.7
+	 *
+	 * @param {(integer|string)}  id       Optimization Problem ID
+	 * @param {jsonschema:Address.Many}   addresses   Address objects
+	 * @param {boolean} reoptimize Determine, whether the `Optimization` should be reoptimized
+	 * @param {module:route4me-node~RequestCallback<jsonschema:Address.Many>}
+	 */
+	linkAddress(id, addresses, reoptimize, callback) {
+		const verror = this.r._validateArgs(arguments, "sdk:Optimizations.linkAddress")
+		if (verror instanceof Error) { return callback(verror) }
+
+		return this.r._makeRequest({
+			method: "PUT",
+			path: "/api.v4/optimization_problem.php",
+			qs: {
+				"optimization_problem_id": id,
+				"reoptimize": reoptimize ? "1" : "0",
+			},
+			body: addresses,
+			schemaName: "Address.Many",
+		}, callback)
+	}
+
+	/**
+	 * Remove a destination (an address) with specified route_destination_id
+	 * from an optimization problem with specified optimization_problem_id.
+	 *
+	 * @see {@link https://route4me.io/docs/#remove-an-address-from-an-optimization Route4Me API}
+	 * @category Optimizations
+	 * @since 0.1.7
+	 *
+	 * @todo TODO: There is no schema for validation an output
+	 *
+	 * @example
+	 * const response = {
+	 *	"deleted":true,
+	 *	"route_destination_id":1
+	 * }
+	 *
+	 * @param {(integer|string)}  id       Optimization Problem ID
+	 * @param {(integer|string)}  routeId  Route destination ID
+	 * @param {module:route4me-node~RequestCallback<jsonschema:Optimizations.UnlinkAddressResponse>}
+	 */
+	unlinkAddress(id, routeId, callback) {
+		const verror = this.r._validateArgs(arguments, "sdk:Optimizations.unlinkAddress")
+		if (verror instanceof Error) { return callback(verror) }
+
+		return this.r._makeRequest({
+			method: "DELETE",
+			path: "/api.v4/address.php",
+			qs: {
+				"optimization_problem_id": id,
+				"route_destination_id": routeId,
+			},
+			schemaName: "Optimizations.UnlinkAddressResponse",
+		}, callback)
+	}
+
 }
 
 module.exports = Optimizations
